[CHORE] : moved type, interface, mock data from Layout.tsx to RoomContext.tsx
